Move server logic to its own file
The code has been refactored to contain the server main logic in the
file `server.ts`, while the `index.ts` will contain only the application
inicialization.
